+++ packages/k2-alpine/src/components/Chart/MiniChart.tsx
@@ -8,7 +8,6 @@ import {
   PaintStyle,
   Path,
   Skia,
-  SkPath,
   StrokeCap,
   TileMode,
   vec
@@ -16,7 +15,7 @@ import {
 import { useMemo, useState } from 'react'
 import { View } from '../Primitives'
 import { useTheme } from '../../hooks'
-import { DataPoint, makeCurve } from '../../utils/chart'
+import { makeCurve } from '../../utils/chart'
 
 export const MiniChart = ({
   style,
@@ -45,19 +44,22 @@ export const MiniChart = ({
     return data.filter((_, index) => index % step === 0)
   }, [data, downsampleTo])
 
-  const { path, lastPoint } = useMemo(
-    () =>
-      makeLineGraph({
-        data: downsampledData.map((item, index) => ({
-          index,
-          value: item.value
-        })),
-        size: graphSize,
-        xPadding: CIRCLE_RADIUS,
-        yPadding: CIRCLE_RADIUS
-      }),
-    [graphSize, downsampledData]
-  )
+  const { path, lastPoint } = useMemo(() => {
+    const curve = makeCurve({
+      data: downsampledData.map((item, index) => ({
+        index,
+        value: item.value
+      })),
+      size: graphSize,
+      xPadding: CIRCLE_RADIUS,
+      yPadding: CIRCLE_RADIUS
+    })
+
+    return {
+      path: curve.path ? Skia.Path.MakeFromSVGString(curve.path) : null,
+      lastPoint: curve.lastPoint
+    }
+  }, [graphSize, downsampledData])
 
   const circleColor = negative ? NEGATIVE_COLOR_RIGHT : POSITIVE_COLOR_RIGHT
 
@@ -122,33 +124,6 @@ export const MiniChart = ({
 }
 
 const CIRCLE_RADIUS = 4
-
-const makeLineGraph = ({
-  data,
-  size,
-  xPadding = 0,
-  yPadding = 0
-}: {
-  data: DataPoint[]
-  size: { width: number; height: number }
-  xPadding?: number
-  yPadding?: number
-}): {
-  path: SkPath | null
-  lastPoint: { x: number; y: number } | null
-} => {
-  const { path, lastPoint } = makeCurve({
-    data,
-    size,
-    xPadding,
-    yPadding
-  })
-
-  return {
-    path: path ? Skia.Path.MakeFromSVGString(path) : null,
-    lastPoint
-  }
-}
 
 const NEGATIVE_COLOR_LEFT = '#EA4542'
 const NEGATIVE_COLOR_RIGHT = '#F5643B'
